Build version 406510, abnorm fixes and changes, added new monsters

+++ HunterPie/Core/Monster/MonsterData.cs
@@ -35,9 +35,9 @@ namespace HunterPie.Core {
             float Mini = float.Parse(Crowns.Attributes["Mini"].Value);
             float Silver = float.Parse(Crowns.Attributes["Silver"].Value);
             float Gold = float.Parse(Crowns.Attributes["Gold"].Value);
-            if (multiplier >= Gold) return "CROWN_GOLD";
-            if (multiplier >= Silver) return "CROWN_SILVER";
-            if (multiplier <= Mini) return "CROWN_MINI";
+            if ((int)multiplier >= Gold) return "CROWN_GOLD";
+            if ((int)multiplier >= Silver) return "CROWN_SILVER";
+            if ((int)multiplier <= Mini) return "CROWN_MINI";
             return null;
         }
     }
